revert change of max file size to long as it will impact index size and 2gb is a reasonable max atm
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@802781 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/store/kahadb/KahaDBPersistenceAdapter.java
@@ -193,7 +193,7 @@ public class KahaDBPersistenceAdapter implements PersistenceAdapter {
      * Get the journalMaxFileLength
      * @return the journalMaxFileLength
      */
-    public long getJournalMaxFileLength() {
+    public int getJournalMaxFileLength() {
         return this.letter.getJournalMaxFileLength();
     }
 
@@ -204,7 +204,7 @@ public class KahaDBPersistenceAdapter implements PersistenceAdapter {
      * 
      * @org.apache.xbean.Property propertyEditor="org.apache.activemq.util.MemoryPropertyEditor"
      */
-    public void setJournalMaxFileLength(long journalMaxFileLength) {
+    public void setJournalMaxFileLength(int journalMaxFileLength) {
         this.letter.setJournalMaxFileLength(journalMaxFileLength);
     }
